+++ cmd/cortextool/CHANGELOG.md
@@ -4,6 +4,7 @@ Order should be `CHANGE`, `FEATURE`, `ENHANCEMENT`, and `BUGFIX`
 
 ## unreleased/master
 
+* [FEATURE] Add `--label-excluded-rule-groups` support to `cortextool rules prepare` command. #174
 * [ENHANCEMENT] Upgrade the Go version used in build images and tests to golang 1.16.3 to match upstream Cortex. #165
 
 ## v0.9.0